les boucles

+++ langage c variable/challenge2.C
@@ -1,9 +1,9 @@
 int main() {
-int F,C;
+float F,C;
 printf("entrer la temperature C:");
-scanf("%d",&C);
+scanf("%f",&C);
 F=(C*1.8)+32;
-printf("la temperature en fahrenheit est : %d",F);
+printf("la temperature en fahrenheit est : %.2f",F);
 
 	return 0;
 }
